ActiveBits

+++ activebits.go
@@ -1,10 +1,10 @@
 package piscine
 
-func ActiveBits(n int) int {
+func ActiveBits(n int) uint {
 
 	binary := ConvertFromDecimal(n, "01")
 
 	result := CountInstances(binary, "1")
 
-	return result
+	return uint(result)
 }
